requirements.txt
discord
PyGithub